Implemented simple floating for boxes/containers.
This means floating works well if there are just one or two frames on the same y position/height.
What does not work well is the case that there are too much frames and they do not fit in on the
same 'line' so that some would need to be wrapped below the other frames. In that case frames will overlap.

This makes the wrap example page look quite ugly regarding frames.
But it works well if you have got just a text paragraph and e.g. one frame to include a picture
or table which shall be surrounded by the text of the paragraph.