feat: add pdf.less for export with dw2pdf
dw2pdf uses the print.css if there is no pdf.css available. It also
allows the usage of the style.css for configured plugins. However wrap's
print.css removes the left margin from note boxes and causes the text to
be displayed above the icon in the resulting pdf.

Adding a pdf.less fixes that problem and allows for feature changes
affecting pdf only.