fix: account for startSectionEdit signature change  (backwards compatible)
In splitbrain/dokuwiki#2220 the signature of \Doku_Renderer_xhtml::startSectionEdit
was changed in a way that breaks backwards compatibility.

In a previous post-Frusterick Manners PR, splitbrain/dokuwiki#2115, the
constant SEC_EDIT_PATTERN was introduced. This allows us to use this
constant to differentiate between Frusterick Manners and Greebo releases
and use the respective argument types.

This added complexity can be removed when Frusterick Manners is no
longer supported by the wrap plugin or DokuWiki core is adjusted to be
backwards compatible.